refactor: Перенёс NewUserRequest, UserDtoResponse, UserShortDto
