+++ ewm-main-service/src/main/java/ru/practicum/dto/event/EventFullDtoResponse.java
@@ -6,7 +6,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import ru.practicum.category.dto.CategoryDtoResponse;
-import ru.practicum.dto.user.UserShortDto;
+import ru.practicum.user.dto.UserShortDto;
 import ru.practicum.json.CustomLocalDateTimeSerializer;
 
 import java.time.LocalDateTime;
