+++ ewm-main-service/src/main/java/ru/practicum/user/controller/AdminUserController.java
@@ -7,7 +7,7 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
-import ru.practicum.dto.user.NewUserRequest;
+import ru.practicum.user.dto.NewUserRequest;
 import ru.practicum.user.service.AdminUserService;
 
 import javax.validation.Valid;
