+++ ewm-main-service/src/main/java/ru/practicum/user/dto/UserShortDto.java
@@ -1,4 +1,4 @@
-package ru.practicum.dto.user;
+package ru.practicum.user.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -6,21 +6,16 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 /**
- * Dto для сущности пользователь. Используется в качестве ответа.
+ * Dto пользователя.
  *
  * @author Nikolay Radzivon
- * @Date 22.06.2024
+ * @Date 19.06.2024
  */
-@Data
+@AllArgsConstructor
 @Builder
 @NoArgsConstructor
-@AllArgsConstructor
-public class UserDtoResponse {
-    /**
-     * Электронная почта пользователя.
-     */
-    private String email;
-
+@Data
+public class UserShortDto {
     /**
      * Идентификационный номер пользователя.
      */
