+++ ewm-main-service/src/main/java/ru/practicum/user/mapper/UserMapper.java
@@ -1,8 +1,8 @@
 package ru.practicum.user.mapper;
 
 import org.springframework.stereotype.Component;
-import ru.practicum.dto.user.NewUserRequest;
-import ru.practicum.dto.user.UserDtoResponse;
+import ru.practicum.user.dto.NewUserRequest;
+import ru.practicum.user.dto.UserDtoResponse;
 import ru.practicum.user.model.User;
 
 /**
