+++ ewm-main-service/src/main/java/ru/practicum/user/service/impl/AdminUserServiceImpl.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
-import ru.practicum.dto.user.NewUserRequest;
-import ru.practicum.dto.user.UserDtoResponse;
+import ru.practicum.user.dto.NewUserRequest;
+import ru.practicum.user.dto.UserDtoResponse;
 import ru.practicum.user.mapper.UserMapper;
 import ru.practicum.user.repository.UserRepository;
 import ru.practicum.user.service.AdminUserService;
