+++ ewm-main-service/src/test/java/ru/practicum/user/controller/AdminUserControllerTest.java
@@ -11,8 +11,8 @@ import org.springframework.http.MediaType;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
-import ru.practicum.dto.user.NewUserRequest;
-import ru.practicum.dto.user.UserDtoResponse;
+import ru.practicum.user.dto.NewUserRequest;
+import ru.practicum.user.dto.UserDtoResponse;
 import ru.practicum.exception.NotFoundUserException;
 import ru.practicum.user.service.AdminUserService;
